Progress on new categoryVolumes code.  Passes tests up to element 44 of 5colorskel.log.
Instead of always creating SimpleIntersections, find_one_intersection
and find_two_intersections in HomogeneityTet return either a
SimpleIntersection or a MultiFaceIntersection (when one of the
orthogonal pixel planes is also a face).  Doing this wrong was
messing up coincidence resolution and pixel plane facet loop
completion.

PixelPlaneFacet::completeLoops ignores coincidences of size 1.
I'm not sure why this wasn't a problem earlier, but the change above
required changing the way that possible coincidences were stored, and
maybe the trivial cases were weeded out.

The case of two antiparallel segments that intersect a polygon at
a point is handled correctly by the coincidence checker.

The relative position of a point along a polygon segment is computed
on-demand and cached (by getPolyFrac), instead of up front in the
coincidence checker.

Changed PixelPlaneIntersection::onSamePolySegment to onOnePolySegment.
It now returns true only if the two intersection points have exactly
one common polygon segment, instead of one or more.