Better coincidence checking for face facets and better area calculation.
Added a step to the coincidence checking for intersections on tet face
edges.  Points that are all at the same parametric position along the
edge are eliminated in start/stop pairs before checking for misordering
due to roundoff.

Fixed the area calculation for pixel plane facets.  Sometimes facets
just consist of pairs of mutually opposed segments, and their area
should be zero.  It's important that it be exactly zero, because that's
how the code knows to check for a completely homogeneous face.
PixelPlaneFacet::getArea now checks for opposing segments and doesn't
include them.