Added a stray pointer so I'm committing and changing platforms to run valgrind.
Added experimental code that includes some otherwise trivial faces
in the computation, because that will generate some infinitesimal
edges that are needed to connect to other infinitesimal edges
already being generated on other faces.

Fixed a bug in PixelPlaneFacet::getFacePlane, which didn't know that
the planes listed in the tet intersection points could be unoriented.