Fixed a handful of bugs in PixelPlaneFacet and HomogeneityTet.
Added somewhat kludgy functions to reset the intersection equivalence
maps in a HomogeneityTet before going on to the next category.

Changed PixelPlaneFacet::completeLoops so that non-crossing
intersections are not passed to the coincidence checker.  The
coincidence resolving routines are not supposed to work on and are
unnecessary for non-crossing intersections.

Changed a "<" to "<=" in PixelPlaneFacet::addEdges, so that
intersections at the same location are joined by a zero length
segment instead of going all the way around the element.  For some
reason this bug only surfaced after non-crossing intersections
were removed from the coincidence checking.

Marked coincidences.h as OBSOLETE.