Fixed bugs related to coincidence detection for intersections on PixelPlanes.
Rewrote classifyVSBcorner and related methods (PlaneIntersection::getOrdering, etc)
so that they don't assume that VSB facet corners are formed from segments on the
same facet.  It's possible for a VSB facet segment to tee into a VSB cross section
segment.

NEXT: Restructure the way that PlaneIntersections are created and destroyed,
so that equivalence information is preserved from one PixelPlane to another.
Currently the code doesn't notice if one PixelPlane uses a PlaneIntersection
that was already used in another, and equivalences are missed.  This is causing
an error in 5colorskel element 195.  This is an improvement because the previous
version was failing at element 119.