The not-so-new-anymore PlaneIntersection scheme for locating intersections now runs through the 5color test successfully.
Changed checkEquiv so that it compares a single intersection against
all previously found equivalence classes.  Added a TODO about how
this probably isn't the best way of doing it.

Added a virtual HPlane method to check if a plane is in an equivalence class.
This is used by checkEquiv, making checkEquiv possibly slow.