HomogeneityTet keeps track of collinear planes.
The HomogeneityTet constructor now finds the sets of pixel planes
and face planes that are collinear, meaning three or more planes that
intersect in a line.  Currently it only does this when two faces are
collinear with one or more pixel planes, but that might have to change.

HomogeneityTet::findFaceFacets makes sure that all collinear planes
are included in each intersection before matching start and end
points of segments.  This allows some otherwise undetectable matches
to be found.