Fixed more bugs.
PixelPlaneFacets keep track of whether any edges were added on their
perimeter when loops were closed.  If so, then the area can't be
negative, so if it appears to be negative it's due to round-off
error and not because the facet has a hole it it.  Setting the area
to 0.0 fixes a bug in which the entire perimeter was added when the
facet consisted only of an infinitesimal triangle where a polygon
corner was grazed by a voxel set boundary.

PlaneIntersection::onOnePolySegment wasn't handling collinear
planes properly.

SingleFaceMixin::getPolyFrac and
PixelPlaneIntersectionNR::locateOnPolygonEdge were making inconsistent
assumptions about which face to use when there were choices, which
was leading to sorting errors completeLoops.  Now SingleFaceMixin
doesn't cache polyFrac.  It's always recomputed using the edge
number passed in to getPolyFrac.

Changed HomogeneityTet::edgeCoord so that it takes an edge number
argument instead of a face pointer, since that's what it's caller
now knows.