Fixes to FaceFacet code.
FaceFacet::addFaceEdges now uses HomogeneityTet::checkEquiv on
the TripleFaceIntersections that it creates, so that equivalences
with previously created intersections can be detected.

HomogeneityTet::findFaceFacets removes equal and opposite edges
from facets before computing the area.  If the only edges on a
facet are equal and opposite, the area should be zero, but if round-off
makes it negative the entire perimeter could be included erroneously.
PixelPlaneFacets already had this check.