More bug fixes for cases with VSB segments collinear with tet edges and for coincidence checking.
HomogeneityTet::find_one_intersection doesn't consider face planes
that are collinear with the VSB segment, unless no other intersection
is found.  In that case the segment must lie in a tet face.

SimpleIntersections and MultiFaceIntersections can be created without
specifying a tet face, in which case the point location and the
intersecting plane must be provided later.

resolveTwofoldCoincidence takes a crossing count > 1 or < 1 as
a sign of a badly formed crossing, and merges the points.

Fixed PixelPlaneIntersection::onOnePolySegment, which wasn't handling
collinear planes properly.

The tet intersection polygon is written to tetpts-<plane>.lines
in debug mode.