isEquiv_() now uses the HomogeneityTet's collinear plane data when possible.
It uses Plane::nonDegenerate() otherwise.  nonDegenerate is susceptible
to round-off error.  This fixes a problem in which the round-off error
was causing non-equivalent points to appear to be equivalent (because
collinear planes appeared to be different).  The code in isEquiv_ is
really ugly, but serves to demonstrate that the fix works.