Added a new "occupied edge" check to prevent extra segments from being added in findFaceFacets.
Got rid FaceFacet::begin() and FaceFacet::end(), which returned
iterators into the set of edges.  Replaced them with
FaceFacetEdgeSet::edges(), which returns a const reference to the
set.  FaceFacetEdgeSet is a new typedef.

Changed PixelPlaneIntersectionNR::sharedFace(PixelPlaneIntersectionNR*)
to a virtual function and extended it to the entire PlaneIntersection
hierarchy with double dispatch.