IsecEquivalence classes are merged at more points during face facet coincidence checking.
Doing this is probably correct, but it moves the problems around without solving them.
The real problem is that the coincidence checking at the edges of face facets needs to
be planned more carefully.  There are a very limited number of possible configurations,
and those need to be checked for, instead of just checking the validity of adjacent points
in the LooseEnds maps.  That adjacency may be determined by round off error.