More fixes to IntersectionGroup::fixCrossings
The actual crossing check (in FaceEdgeIntersection::crosses) is now
done the way it used to be in findFaceFacets, just using cross and
dot products, and not computing the position of the intersection.
This method explicitly uses the positions of the intersections on
the facet edge, which means that it uses more topological information
and is less susceptible to round off error than the method it
replaces.

Also fixed some bookkeeping errors in other parts of the routine.