Added another check to HomogeneityTet::resolveCoincidences.
The new "fixTents" test picks up misordered points that arise when
the edge of a tet face grazes a voxel corner, leading to two
intersections on the face that are joined by segments that meet
at the voxel corner.  Because it's just a grazing intersection,
the intersection points can be in the wrong order.  Because the
geometry of the voxel corner is known, the incorrect order can
be detected.