Fixed some problems with face facets.
When intersection points are forced onto the closest tet edge because
they're known to be on an edge but aren't, the other face of the
edge is added to the intersection's equivalence class.

Added command line option --verboseWait=n, or --vW=n, which can be used
to skip the first n instances when verbose mode would be turned on
in CSkeletonElement::categoryVolumes.   This is useful when a problem
only appears after a number of iterations of a skeleton modifier.

Added a TODO about making the planes and faces in the IsecEquivalence
class the principle storage for planes and faces for all types of
PlaneIntersection.