Still working on equivalence classes the primary way of comparing intersections.
This code does not work, but I'm committing it now to transfer changes
from the travel laptop.

PlaneIntersection::isEquivalent is no longer a virtual function.
It just compares equivalence classes.

Only the unoriented planes are inserted into the equivalence
classes, but the intersections store the oriented planes.

Changed the names of some functions for clarity.