New HomogeneityTet routines to get the properly oriented orthogonal plane when constructing PlaneIntersections.
When constructing the PlaneIntersections for a VSB loop on a
PixelPlane, previous versions used planes orthogonal to the pixel
plane oriented with their normals outward to the loop.  This was
incorrect.  The new version puts the normals outward to the voxel
category, unless the plane is also a tet face, in which case it
uses the outward normal to the face.

HomogeneityTet::coincidentFacePlanes now maps both orientations of
the coincident pixel planes to a face.

Rewrote HomogeneityTet::find_one_intersection and find_two_intersections
to use the new HomogeneityTet::orientedOrthogonalPlane routines,
and modified doFindPixelPlaneFacets accordingly.  doFindPixelPlaneFacets
no longer assumes that the orthogonal plane along a VSB segment has
the same orientation at both ends of the segment.

Added some methods to PixelBdyLoop and PixelBdyLoopSegment to make
the orthogonal plane computation easier.

Added CMicrostructure::containsPixelCoord(Coord) which just computes
whether or not the given point is inside the microstructure.  We'd
had this for ICoords, but not Coords.

Added operator* for double*ICoord -> Coord.