Fixed a bug in argument checking for intersection coincidence resolution.
PixelPlaneFacet::polyVSBCornerCoincidence was using onSameFacePlane to
check that the points it was comparing weren't on the same face.  The
correct condition is that the points share two faces.  Now there's
an onTwoPolySegments routine that checks that.

SimpleIntersection::isMisordered(SimpleIntersection*) now counts
the number of shared faces instead of using onOnePolySegment when
deciding which coincidence checker to use.  Other intersection classes
may need to be fixed too.

Also, PixelPlaneIntersectionNR::sharedFaces takes a second argument
which is a face to exclude from the result (normally the pixel plane,
when it's coincident with a face).

Added a version of HomogeneityTet::areCollinear that takes a FacePlaneSet
argument, for convenience.