Two fixes to MultiFaceIntersection.
Fixed MultiFaceIntersection::interiority, which needs to use < and >
instead of <= and >=.  I had changed the comparisons (I think) when fixing
the cross products, but that was a mistake.

Made MultiFaceFixin<BASE>::getOtherFaceIndex a virtual function.  The
compiler isn't generating it if it's not declared virtual, even though
I don't see why it has to be virtual.

Deleted some old code and commented out debugging lines.