Changed ElementFuncNodePositionIterator::operator+= and added a TODO about doing it correctly.
This function was obviously wrong (helpfully reported as such by
the new clang++ 8.0.0) but also is apparently not used, at least
by the current test suite.  The new version is probably wrong for
non-linear and non-isoparametric elements, of which we currently
have none.  I added a TODO to the function, as well as a comment
in the existing TODO in CSkeletonElement::realElement().