Rewrote PixelPlaneFacet::getEdgesOnFaces()
This is supposed to fix the problems in the old way of figuring out
which PixelPlaneFacet edges are also on FaceFacets.  It's crashing
strangely (in the RedundantIntersection destructor) so I am committing
it to transfer to Linux where maybe valgrind will be useful.

sharedFaces() used to be a non-virtual method in PixelPlaneIntersectionNR.
Now it's pure virtual in PixelPlaneIntersection and defined in
PixelPlaneIntersectionNR and RedundantIntersection.  It's used by the
new version of PixelPlaneFacet::getEdgesOnFaces().

The various virtual versions of FacetEdge::getEdgesOnFaces are no
longer used.