Fixed a few more bugs.
The sign of the cross products in IntersectionGroup::tentCheck
was wrong, which caused some nearby but not identical
intersection points to be merged incorrectly.

IntersectionGroup::fixCrossings needs to call
GenericIntersection::computeLocation before
GenericIntersection::checkEquiv.

FaceEdgeIntersection::crosses now checks that the far ends
of the segment aren't equivalent before checking the precendence
of the near points.

The setLocation methods in PixelPlaneIntersectionNR and GenericIntersection
now use a new fixedLocation method, which adjusts the points location
so that it's exactly on its pixel planes.  GenericIntersection
didn't used to do this, which was causing problems.

Added checks in debug mode to ensure that points added to
IsecEquivalenceClasses aren't too far from one another.  This check
would have picked up the tentCheck cross product error.

Lots of changes to debugging code and comments.