Fixed problems in IntersectionGroup::tentCheck.
tentCheck looks for topology problems when two segments of a face
facet join at a point interior to the tet face and the other ends
of the segments are on the same tet face edge.  It now checks that
the sign of the area of the triangle formed by the segments and
the tet edge is what is expected given the ordering of the
intersection points on the edge.