More fixes to FaceEdgeIntersection::crosses.
It's now legal to check for crosses between two segments that
both start or stop on the face edge.

Parallel segments are handled better.  segIntersection() wasn't setting
alpha or beta when the segments were parallel -- it was just returning
false.  crosses() was using only alpha and beta and ignoring the
return value.  Now segIntersection sets an additional flag to indicate
that it hasn't computed alpha or beta.