Fixes to HomogeneityTet::doFindPixelPlaneFacets and PixelPlaneFacet::getArea.
doFindPixelPlaneFacets now uses CRectangle::intersects_open when
comparing the bounding boxes for a tet and a segment before calling
find_two_intersections.  This makes it consistent with the similar
calculations done in getTetPlaneIntersectionPoints and in
BarycentricCoord::interior.  Without this fix, some intersections
were found in one pixel plane and not in others, and failed to match
up.

The fix to doFindPixelPlaneFacets led to situations in which a
pixel plane facet edge could join two equivalent end points.
PixelPlaneFacet::getArea has to ignore such segments.