Various fixes, motivated by a bug in SimpleIntersection::mergeWith.
SimpleIntersection::mergeWith(SimpleIntersection) only creates a clone
if the merging points are equivalent *and* are on just one polygon segment
and one loop segment.  It used to only check equivalence, but that was
a mistake for intersections where two pixels touched corners.

IntersectionPlanesBase::computeLocation checks that it's only using
independent pixel planes.  A merged intersection can contain a plane
and its oppositely oriented counterpart, but only one of them should
be used.

Similarly, IntersectionPlanes<>::belongsInEqClass makes sure to only
pass independent pixel planes to isEquiv_.

MultiFaceMixIn<>::getPolyFrac checks the point's equivalence class
to find which polygon sides it might be on.  It used to check only
the faces stored in the point itself.

Added IsecEquivalenceClass::containsFacePlane.

Changes to debugging lines (which I should have commented out
before committing this...)