Fixes to triple coincidence resolution.
PixelPlaneFacet::resolveThreeFoldCoincidence was setting the
crossing type of the merged intersections, which was a holdover
from the days before shifting from crossing types to crossing
counts.

PixelPlaneFacet::tripleCoincidence now checks for cases in
which one of its arguments is not derived from SingleVSBbase,
indicating that it's at the vertex of a checkerboard arrangement
of pixels.