Added another check for parallel segments in segIntersection.
The first check (for zero denominator) can fail due to round off
error, and cause the routine to return alpha=0 when alpha should
be very large.  It turns out that the numerator in the expression
is also (theoretically) zero for parallel segments.  Checking the
numerator at least guarantees that we won't return alpha=0, probably.
As long as we don't accidentally get alpha between 0 and 1 we're ok.