Added two checks to HomogeneityTet::find_two_intersections.
find_two_intersections compares segment direction to face normal
when looking for entries and exits.  For an entry, the VSB segment
dotted with the face normal must be negative.  For an exit, the
dot product must be positive.

find_two_intersections also checks that the intersection point is
actually within the face that's being intersected.  Previously, we
relied on checking that the entry alpha was less than the exit
alpha, but this can fail if, for example, the exit alpha is supposed
to be +epsilon on one face, but round-off error puts it at -epsilon,
(meaning that there's no intersection with that face) so a different
face is used with a much larger alpha.

Adding this check actually causes errors earlier in the test (in
annealbug.log) when a VSB segment exits the tet through an edge and
appears not to be on an exit face.  This will be fixed next.

Resurrected averageBary() which is used in find_two_intersections
to compute the barycentric coords of an intersection point from the
barycentric coords of a segment's endpoints and the position of the
intersection relative to them.