Rewrote HomogeneityTet::find_two_intersections.
The new version no longer uses the barycentric coordinate components
to check that the end points of the segment are on opposite sides
of a tet face before computing the intersection with that face.
The old way was prone to round off error, and it only eliminated
cases that can also be eliminated later.

This version does not check that the computed intersection with a
face plane is actually within the face.  I'm not sure that it's
needed.  On the other hand, this version fails too, but later than
the previous versions...