Improved find_one_intersection.
HomogeneityTet::find_one_intersection now only checks the face
planes that bound the tet intersection polygon.  This fixes an
error that only showed up the 46th time element 177 was evaluated
in annealbug.log, so we're making progress!  The bounding faces
are stored in the PixelPlaneFacet by its constructor.  Without
this check, its possible for an intersection to be identified
by a face that's not associated with a polygon edge, which will
cause problems in PixelPlaneIntersectionNR::locateOnPolygonEdge.

The same modification still needs to be made to find_two_intersections.

PixelPlaneIntersectionNR::locateOnPolygonEdge uses planes from the
equivalence class.  This satisfies a TODO, but didn't have a
noticable effect.  It's probably the right thing to do.