Two fixes to pixel plane facet construction.
The argument onFace was being passed to the PixelPlaneFacet
constructor as a boolean, which it used to be, but it's now
supposed to be an int indicating which tet the facet is on.

resolveMultipleCoincidence was failing because it was getting the
wrong value for totalIntersections, because the return value of
storeCoincidenceData hadn't been updated when we started treating
a single intersection as a trivial coincidence.

Rewrote the main loop in resolveMultipleCoincidence to include
improvements that had been made in resolveTwoFoldCoincidence.

'origin/rewriting-category-volumes' by 1 commit.  #   (use "git
push" to publish your local commits) # # Changes to be committed: