Added FaceFacetLoop class and its area() method, which calculates areas more carefully.
If a face facet consists of a small triangle (or triangles) formed
when the corner of a voxel set barely touches the face, then the
area of the triangle can be computed incorrectly if only the edges
are used.  If the area of a face facet is near zero and in danger
of suffering from this problem, the facet is broken up into loops
whose area is computed separately.  FaceFacetLoop::area notices
when the calculation might be wrong and uses topological information
to determine the sign of the area.  The correct sign is important
in fixNonpositiveArea.