Fixed a bug in polyVSBCornerCoincidence.
When checking to see if the VSB corner is inside the polygon,
it should only uses the tet faces that form the edges of the polygon
that the intersections are on.  It's possible for the VSB corner to
be outside the polygon if it's across a third unrelated edge, but this
doesn't affect the topology of the intersections in question.