Fixed a IsecEquivalenceClass bug and a getPixelPlane bug.
After merging two IsecEquivalence classes, HomogeneityTet::mergeEquiv
searches for other equivalence classes that may be equivalent to
the merged one, and merges them too.  This doesn't fix a bug that
was arising in skelrefinebug.log, but it fixes a problem that was
leading to the situation in which the bug was arising, therefore
making the bug unfixable for now.

Added IsecEquivalenceClass::isEquivalent which checks that two
equivalence classes are equivalent.

Fixed a bug in HomogeneityTet::getUnorientedPixelPlane, which
wasn't setting an unoriented plane's unoriented plane pointer
to itself.  This was causing a crash in the new mergeEquiv.