Fixed (I think) a bug in PixelPlaneFacet::badTopology(SimpleIntersection, MultiVSBIntersection)
badTopology was just wrong about the topological constraints on
some of the legal configurations of a SimpleIntersection and a
MultiVSBIntersection at a VSB corner.

This version fails the annealbug test on my laptop but not on
my desktop, which is peculiar because they are using the same
compiler (although different cpus).  I'll test it on Linux next.