Added EdgePosition class to avoid round-off error when comparing relative positions of intersections on edges.
EdgePosition objects are returned by PlaneIntersection::getPolyFrac and
HomogeneityTet::faceEdgeCoord and related functions.  Those methods
are careful to compute relative positions in a way that's independent
of the direction of the edge.  If the direction used for the computation
is different from the direction of the edge, EdgePosition::reversed
is set to true.  Comparisons between reversed positions are made without
subtracting the positions from 1, which can introduce round-off error.