Improved debugging output, and experimental changes to matchStrandedPoints
Added experimental block that shortcuts matchStrandedPoints and
treats all stranded points as marooned, so that findFaceFacets just
moves them to the nearest tet edge.  This revealed problems in the
organization of findFaceFacets.  When points are merged and/or moved
onto an edge, it's necessary to re-run all of findFaceFacets up to
that point, since moving points onto an edge may have the effect
of moving a pixel plane facet edge onto a tet face.