A few fixes and modifications to coincidence detection on tet faces.
This version passes the annealbug.log test and gets farther than
ever before on snaprefinecatbug.log.

Changed HomogeneityTet::findFaceFacets so that it performs its
topological tests on edges and points on all faces, not just the
ones that don't align with pixel planes.  Facets don't have to be
computed again on the aligned faces, but correcting topological
errors on those faces can move points to the tet edges, where they
become important for the neighboring faces.

Changed EXPERIMENTAL to true in HomogeneityTet::matchStrandedPoints,
which means that matchStrandedPoints isn't actually necessary and
will be eliminated soon.

It's possible for the planes found in FaceFacetLoop::area to fail
to be mutually orthogonal, if the loop is entirely due to round off
error where a tet corner touches a VSB plane.  This condition is
now detected and the area set to 0.

Added EdgePosition::forceToEnd() that eliminates round off error
when an intersection point is known to be at one end of a tet edge.
HomogeneityTet::findFaceEdge uses it.  This resolves a problem
in IntersectionGroup::fixCrossings.