A few fixes...
Rewrote HomogeneityTet::checkEquiv so that it handles merging multiple
equivalence classes better.

Changed PixelPlaneFacet::badTopology(SimpleIntersection, MultiVSBIntersection)
so that it doesn't use most of the tests that it used to use, because I found
situations in which they were incorrect.

Fixed IntersectionPlanesBase<>::belongsInEqClass so that it uses the
intersection's equivalence class, if it has one.  This fixes some problems
in which equivalences weren't being detected.

Changes to debugging lines elsewhere.