Added FaceFacet::removePerimeterEdges to clean up before completing the facet.
removePerimeterEdges removes all face facet edges that lie along
the perimeter of the tet face.  These edges can be spurious even
after the coincidence resolution step.  If the edges are real,
they'll be recreated when segments are added to complete the facet
on the perimeter.

This version passes annealbug.log and most of snaprefinecatbug.log.