Various fixes.
FaceEdgeIntersection::crossesSameEdge only reports intersections that
are near the beginning of the segments.  Intersections at the far
end of the segments will be handled at a different coincidence location.
This avoids the situation in which a segment ends on another segment
but appears to cross it.

Permanently removed the minimum area check at the start of
HomogeneityTet::doFindPixelPlaneFacets, and added a comment about why
it's a bad idea.

Added a test for degenerate polygons before doing the winding number test
for homogeneous pixel plane facets at the end of doFindPixelPlaneFacets.

Debugging changes elsewhere.