The physical location of all PlaneIntersection objects is computed before the objects are constructed.
Computing the position before construction ensures that the position
is computable.  Previously it was possible for an intersection to
be constructed from planes that seemed to intersect by one test,
but for which the solution of the matrix equation for the intersection
point failed.  Now, in those cases in which the matrix solution is
necessary, the success of the solution is the test for whether or
not the planes intersect, so everything's consistent.

findIntersectionLocation() in planeintersection.C intelligently
computes the intersection point of sets of PixelPlanes, FacePlanes,
and FacePixelPlanes, and indicates if the intersection exists.
Other routines had to be modified to handle sets of planes instead
of PlaneIntersections.  Perhaps there should be a class that includes
just the three sets, and which is included in a PlaneIntersection.

Moved triplePlaneIntersection() from common/pixelsetboundary.C to
engine/planeintersection.C, where it's static.

Removed some obsolete code.