A couple of unresolved changes related to intersections of single tet nodes on pixel planes.
Changed getTetPlaneIntersectionPoints so that single tet nodes on the plane
are treated as intersections.  This leads to a degenerate intersection polygon
on the plane, which allows degenerate facet edges to be created on neighboring
faces, which should fix problems with loose ends from other segments on those
faces.  It also causes other problems, namely when a single point intersection
is on the edge of a voxel set on a pixel plane, and the degenerate intersection
polygon creates facet edges on faces that shouldn't have facets at all.

find_two_intersections checks to see if the given segment passes through
a FacePixelPlane, even if the segment is nominally outside the bounding
box, because the intersection with the FacePixelPlane might be at a node
that's within the tet, by some round-off error induced nonconvexity.
This didn't actually solve the problem it was meant to solve, because the
case was never considered, because getTetPlaneInteresctionPoints wasn't
treating single points as intersections.

Other changes to comments and debugging code.

At this point I'm abandoning this branch to try the r3d scheme of Powell and Abel
instead.