Code to create the graph-based VoxelSetBoundary appears to be working.
The VSBs are created properly for the 5color microstructure, at least
according to visual inspection.  More rigorous tests still need to be
created, as well as the clipping algorithm and homogeneity calculation.

Commented out some old code for displaying and debugging the old
VoxelSetBoundary class and associated machinery.  Many of the files
now contain "OBSOLETE" to ensure that they're not compiled.  They'll
be removed soon.

checkSignatures.py is a script that checks that the 2x2x2 voxel signatures
defined in voxelsetboundary.C are unique.  I didn't mean to commit it.
I'll move it somewhere else.