Debugging the code for generating the VSBGraph from the voxels.
Added a new file to the regression suite for testing the VSBGraph
and its clipping operations.
The All2x2x2 test generates all 256 configurations of voxels
in a 2x2x2 cube to check that the graph edges are connected
correctly.  This version is segfaulting on the checkerboard
configuration.  I'm committing this now to move it to Linux to run
valgrind.

Added a --noclean option to regression.py, to keep it from removing its
temp directory.