VSBGraph::copyAndClip seems to be working.
copyAndClip creates a new VSBGraph that is clipped by one plane.
It's the first step in clipping a graph by a tetrahedron.  The
remaining steps will be done without copying the graph.

Added COrientedPlane class in geometry.*.

Added reference files for the graphs of all of the 2x2x2 voxel configurations.
voxel_test.py checks them.