New tests for the homogeneity calculation.
CSkeleton::checkCategoryVolumes checks that the sum of the volumes
of a voxel category in each element is the same as the total
volume of the category as determined by counting the voxels in
the category.

Added a bool argument to categoryVolumes to tell it when to run the
o(N^3) connectivity tests.   This is only set by checkCategoryVolumes,
and the check is only run on completely clipped VSBs, to keep N
small.

Changed the return type of CSkeletonElement::categoryVolumes from
std::vector<double>* to std::vector<double>.  With C+11 move
constructors, there's no need to return a pointer.

Removed some debugging lines and other old code.