Fixed a problem in the new homogeneity calculation.
It has to use the total element volume computed from the sum of the
volumes of each category, and not from the positions of the nodes.
Otherwise, round off error may make an element with only one category
have a homogeneity slightly less than 1.0, and operations with a
minimum homogeneity threshold of 1.0 will have surprising results.