Various changes for checking the new categoryVolumes routines.
Added CSkeleton::compare2, which compares two skeletons to see if
they differ only by the choice of the added diagonal in
Refine::tet6Edges1Div.  This choice can be affected by roundoff
error in the homogeneity calculation, and was making the first set
of refinement tests in skeleton_modification_test.py fail in
uninteresting and unimportant ways.

Added Coord3D::operator=.  Sort of surprised that it didn't already exist.

Just added a TODO to crefine.C.