Added a display method for displaying segments that are in one Skeleton but not in another.
This is for debugging -- checking that the differences between the
skeleton_modification_test output file and refereence files are
understandable.  I started to try to create a SkeletonFilter that
would just select the appropriate segments, but that started to get
messy.  Doing it right will require new kinds of GridSource subclasses
that each have their own type of SkeletonFilter.  One DisplayMethod
may need more than one GridSource and vtk pipeline.  It would make
more sense to do this *after* upgrading to vtk 6 or 7.  So the code added
here is sort of ugly and is all inside #ifdef DEBUG blocks.