Minor changes
Added routines to save a clipped VSB to a file, for debugging.
The previous routine would only save after clipping a single plane.
The new one handles more than one clipping plane.

Clipped VSBs are saved using phsyical coordinates, so they can be
plotted with the Skeleton even if the voxel size isn't 1x1x1.

CSkeleton::calculateHomogeneityIndex uses the sum of the volumes
of the categories instead of the geometrically computed element
volume in the divisor.  Because the element volume is computed
differently from the category volumes, round-off error could make
the homogeneity index be greater than 1.  This is really just a
cosmetic problem.

Changed VOLTOL to 1.e-6, since the category volume calculation seems
to be off by that much in some otherwise completely unremarkable
elements.

Added a TODO about improving VSBGraph::getDistances().