Support for the profiler from google perftools.
Building the profile3d version in oof2-build #defines PROFILER in C++.
In Python, config.profiler() indicates whether profiling is turned on
or not.

cdebug contains swigged functions for starting and stopping the profiler.

Use ProfilerStart(filename) and ProfilerStop() to surround the code to
be profiled, in either C++ or Python.