Minor changes to get the regression tests to run on the 3Dwidget branch.
I'm not 100% sure that these changes are correct... The regression
tests run in non-GUI mode, and the viewer tool box had methods that
refer to the current clipping plane, but currentClipPlane() was
only defined in viewertoolboxGUI.   I added a method that returns
None in the non-gui part of the toolbox.

The material display in microstructuredisplay.py was not setting
opacity.  Bitmap displays in the 3Dwidget branch have an opacity
parameter.  I don't remember if opacity was removed from displays
on other branches after the 3Dwidget branch was started, or added
to the displays on the 3Dwidget branch.  For now, I just added an
opacity parameter to the material display.  It may need to be
removed.  It doesn't seem to do anything.