Fixed problems with CSubProblem::get_mesh, set_mesh, and mesh in the python wrapper class.
This mostly reverts to the old version in master.  Some changes
made on the category volumes branches were incorrect and confused
and somehow got past the checks and onto the development branch.

The confusion is about how the CSubProblems are to know about
their FEMesh on the Python side.  There were conflicting comments
about how it's supposed to work, and some half-baked code.

* FEMesh.__init__ was hacked in femesh.spy to add some data to
the python wrapper class, which was justified by saying that an
FEMesh is never returned from C++ to Python, so adding the data to
one particular instance of the wrapper was ok.

* Some versions of csubproblem.* defined python functions for
setting and getting the FEmesh in the python wrapper
for CSubProblem.  This was so that the FEMesh didn't have to be
returned from C++ to python (and re-wrapped, therefore losing the
extra data added in FEMesh.__init__.

* Other versions of csubproblem.* had comments about getting rid of
the mesh stored in the wrapped CSubProblem, and used swigged
get_mesh and set_mesh methods.  (Except that set_mesh was called
set_femesh sometimes.)   This contradicted the comments in femesh.spy
about not returning FEMesh's from C++ to Python.

Now I've restored CSubProblem get_mesh and set_mesh in swig, deleted
CSubProblem.mesh in python, and added comments about how FEMesh
should *not* redefine __init__ and should store the extra data in
MeshContext instead.