Working on fixing bugs in the clipping plane machinery.
I'm committing now just to be able to run valgrind on linux. This
version has a lot of debugging code added.  It crashes when planes
are manipulated while one plane is selected in the toolbox.

The fundamental problem is that the selection state for clipping
planes was stored only in the GUI, so test scripts that use clipping
planes failed in text mode.  Now the current plane is stored in
the non-gui part of the ViewerToolbox where it is manipulated by
menu commands which issue switchboard signals to update the GUI.
But something is screwy with the construction/destruction of clipping
planes, or something like that.

The toolboxes' "active" status was serving two roles, which were
conflicting.  It was indicating both that the toolbox was the current
toolbox, and that the toolbox's mouse handler was installed.  Using
the Toggle button on the graphics window would mark the toolbox as
inactive while it was still current.  Now "active" only means that
the toolbox is current and says nothing about the mouse handler.
The toolboxes have a new installMouseHandler method that is called
when their mouse handler has to be installed.

Added lots of setModified() calls to PlaneAndArrowLayer methods,
because the display wasn't updating properly.  setModified()
calls the Modified() methods of the arrow and plane sources.

Swigged the ClippingPlane equality operator so that planes in
the GUI list can be selected by value, not index.  This hasn't
been done yet, though.

Added summary.txt, which must have been written to get Ryan
started, and might not be obsolete yet.  I'm not sure why
it's showing up as a new file now, though.