Fixed a ClippingPlane/View bug.
Views returned to Python by GhostOOFCanvas::get_view were %new'd
in swig, so Python calls the View destructor when they go out of
scope.  But View::getClipPlane returns a reference to a ClippingPlane
stored within the View, which can outlive the View in the calling
Python code.  Now the ClippingPlanes are cloned when they might
outlive their Views.