Fixed some problems with the clipping plane widget and the viewer toolbox.
Reworked a lot of the switchboard communication between the toolbox,
the gui toolbox, and the plane-and-arrow widget.  The "clip selection
changed" signal is now used for more situations.  There's no "clip
edited", "clip inversion changed", or "clip suppresion changed"
anymore.

The non-gui toolbox now stores the current clip plane, so that
menu commands will work in text mode.   The gui toolbox doesn't
store any data outside of the widgets.

Refactored the methods in the GUI toolbox to reduce the side
effects of each routine.

Clip planes can now be selected and displayed even if they're not
enabled.

New clip planes are automatically selected.  The selected plane
and the displayed plane agree.